fix: внести изменения по результатам ревью и произвести рефакторинг.
