+++ ewm_service/src/main/java/ru/practicum/ewm/admin_api/compilations/CompilationsService.java
@@ -1,15 +1,16 @@
-package ru.practicum.ewm.public_api.compilations;
+package ru.practicum.ewm.admin_api.compilations;
 
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
-import ru.practicum.ewm.admin_api.compilations.CompilationsMapper;
-import ru.practicum.ewm.admin_api.compilations.CompilationsRepository;
 import ru.practicum.ewm.admin_api.compilations.model.Compilation;
 import ru.practicum.ewm.admin_api.compilations.model.CompilationDto;
+import ru.practicum.ewm.admin_api.compilations.model.NewCompilationDto;
+import ru.practicum.ewm.admin_api.compilations.model.UpdateCompilationRequest;
 import ru.practicum.ewm.admin_api.events.EventsMapper;
 import ru.practicum.ewm.admin_api.events.EventsRepository;
+import ru.practicum.ewm.admin_api.events.model.Event;
 import ru.practicum.ewm.admin_api.events.model.EventShortDto;
 import ru.practicum.ewm.exeptions.NotFoundException;
 
@@ -24,6 +25,42 @@ public class CompilationsService {
     private final EventsRepository eventsRepository;
     private final CompilationsMapper compilationsMapper;
     private final EventsMapper eventsMapper;
+
+    public CompilationDto create(NewCompilationDto newCompilationDto) {
+        Compilation compilation = compilationsMapper.toCompilation(newCompilationDto);
+        List<Event> events = new ArrayList<>();
+        if (newCompilationDto.getEvents() != null) {
+            events = eventsRepository.findAllById(newCompilationDto.getEvents());
+            compilation.setEvents(events);
+        }
+        CompilationDto compilationDto = compilationsMapper.toCompilationDto(compilationsRepository.save(compilation));
+        if (newCompilationDto.getEvents() != null) {
+            compilationDto.setEvents(events.stream()
+                    .map(eventsMapper::toEventShortDto)
+                    .collect(Collectors.toList()));
+        }
+        return compilationDto;
+    }
+
+    public CompilationDto update(long id, UpdateCompilationRequest updateCompilationRequest) {
+        List<Event> events = new ArrayList<>();
+        Compilation compilation = compilationsRepository.findById(id).orElseThrow(() -> new NotFoundException("Compilation not found: " + id));
+        if (updateCompilationRequest.getEvents() != null) {
+            events = eventsRepository.findAllById(updateCompilationRequest.getEvents());
+            compilation.setEvents(events);
+        }
+        if (updateCompilationRequest.getPinned() != null) compilation.setPinned(updateCompilationRequest.getPinned());
+        if (updateCompilationRequest.getTitle() != null) compilation.setTitle(updateCompilationRequest.getTitle());
+        CompilationDto compilationDto = compilationsMapper.toCompilationDto(compilationsRepository.save(compilation));
+        compilationDto.setEvents(events.stream()
+                .map(eventsMapper::toEventShortDto)
+                .collect(Collectors.toList()));
+        return compilationDto;
+    }
+
+    public void delete(long id) {
+        compilationsRepository.deleteById(id);
+    }
 
     public CompilationDto getById(long id) {
         Compilation compilation = compilationsRepository.findById(id)
@@ -47,14 +84,6 @@ public class CompilationsService {
             compilationDtos.add(compilationDto);
         }
         return compilationDtos;
-        /*  return compilations.stream()
-                .map(compilation -> {
-                    CompilationDto compilationDto = compilationsMapper.toCompilationDto(compilation);
-                    List<EventShortDto> events = getEvents(compilation.getEvents());
-                    compilationDto.setEvents(events);
-                    return compilationDto;
-                })
-                .collect(Collectors.toList());*/
     }
 
     private List<EventShortDto> getEvents(List<Long> ids) {
@@ -62,5 +91,4 @@ public class CompilationsService {
                 .map(eventsMapper::toEventShortDto)
                 .collect(Collectors.toList());
     }
-
 }
