+++ ewm_service/src/main/java/ru/practicum/ewm/admin_api/users/UserControllerAdmin.java
@@ -24,7 +24,7 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(path = "/admin/users")
-public class UserController {
+public class UserControllerAdmin {
 
     private final UserService userService;
