+++ stats_service/stats_dto/src/main/java/ru/practicum/ewm/EndpointHit.java
@@ -27,6 +27,7 @@ public class EndpointHit {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
     private int id;
     private String app;
     private String uri;
